Updated file reading mechanism to read qvh and ios_video_pull format
THe file mechanism was expecting a file format that was a hack of qvh to contain nalu size before each nalu.
Updated the code to be able to process the standard upstream output of qvh, and also able to process the version written by ios_video_pull containing json info blocks before each nalu.